+++ afkak/test/test_client.py
@@ -16,6 +16,9 @@ from twisted.internet.defer import (
     )
 from twisted.internet.error import ConnectionRefusedError
 from twisted.test.proto_helpers import MemoryReactorClock
+from twisted.names import dns
+from twisted.names.dns import RRHeader, Record_A, Record_CNAME
+from twisted.names.error import DNSNameError
 
 import struct
 import logging
@@ -37,6 +40,7 @@ from afkak.common import (
 )
 from afkak.kafkacodec import (create_message, KafkaCodec)
 from afkak.client import _collect_hosts, _get_IP_addresses
+import afkak.client as kclient  # for patching
 
 DEBUGGING = True
 setDebugging(DEBUGGING)
@@ -154,21 +158,53 @@ class TestKafkaClient(unittest.TestCase):
         mocked_brokers = {
             ('kafka21', 9092): MagicMock(),
             ('kafka22', 9092): MagicMock(),
-            ('kafka23', 9092): MagicMock()
         }
         # inject broker side effects
-        mocked_brokers[('kafka21', 9092)].makeRequest.side_effect = \
-            RequestTimedOutError("kafka01 went away (unittest)")
+        mocked_brokers[
+            ('kafka21', 9092)
+            ].makeRequest.side_effect = lambda a, b: fail(
+            RequestTimedOutError("Kafka21 went away (unittest)"))
         mocked_brokers[('kafka22', 9092)].makeRequest.return_value = \
             succeed('valid response')
-        mocked_brokers[('kafka23', 9092)].makeRequest.side_effect = \
-            RequestTimedOutError("kafka03 went away (unittest)")
 
         client = KafkaClient(hosts='kafka21:9092,kafka22:9092')
 
         # Alter the client's brokerclient dict
         client.clients = mocked_brokers
         client._collect_hosts_d = None
+        resp = self.successResultOf(
+            client._send_broker_unaware_request(1, 'fake request'))
+
+        self.assertEqual('valid response', resp)
+        mocked_brokers[
+            ('kafka22', 9092)].makeRequest.assert_called_with(
+                1, 'fake request')
+
+    @patch('afkak.client._collect_hosts')
+    def test_send_broker_unaware_request_reresolve_fails(self, _collect_hosts):
+        """
+        test_send_broker_unaware_request_reresolve_fails
+        Tests that call works when at least one of the host is available, even
+        if we attempted to re-resolve our hostnames and no IPs were returned.
+        """
+        mocked_brokers = {
+            ('kafka21', 9092): MagicMock(),
+            ('kafka22', 9092): MagicMock(),
+        }
+        # inject broker side effects
+        mocked_brokers[
+            ('kafka21', 9092)
+            ].makeRequest.side_effect = lambda a, b: fail(
+            RequestTimedOutError("Kafka21 went away (unittest)"))
+        mocked_brokers[('kafka22', 9092)].makeRequest.return_value = \
+            succeed('valid response')
+
+        client = KafkaClient(hosts='kafka21:9092,kafka22:9092')
+
+        # Alter the client's brokerclient dict
+        client.clients = mocked_brokers
+        client._collect_hosts_d = True
+        _collect_hosts.return_value = succeed([])
         resp = self.successResultOf(
             client._send_broker_unaware_request(1, 'fake request'))
 
@@ -204,6 +240,43 @@ class TestKafkaClient(unittest.TestCase):
         client._collect_hosts_d = None
         respD = client._send_broker_unaware_request(1, 'fake request')
         reactor.advance(client.timeout + 1)  # fire the timeout errback
+        self.successResultOf(
+            self.failUnlessFailure(respD, KafkaUnavailableError))
+        self.assertTrue(cbArg[0].check(RequestTimedOutError))
+
+    @patch('afkak.client.time')
+    def test_make_request_to_broker_alerts_when_blocked(self, time):
+        """test_make_request_to_broker_alerts_when_blocked
+        Test that a blocked reactor will cause an error to be logged.
+        """
+        cbArg = []
+
+        def _recordCallback(_):
+            # Record how the deferred returned by our mocked broker is called
+            cbArg.append(_)
+            return _
+
+        d = Deferred().addBoth(_recordCallback)
+        m = Mock()
+        mocked_brokers = {('kafka31', 9092): m}
+        # inject broker side effects
+        m.makeRequest.return_value = d
+        m.cancelRequest.side_effect = lambda rId, reason: d.errback(reason)
+        m.configure_mock(host='kafka31', port=9092)
+        # And the mock time.time's values
+        time.time.side_effect = [0.0, 10.0, 20.0]
+
+        reactor = MemoryReactorClock()
+        client = KafkaClient(hosts='kafka31:9092', reactor=reactor)
+
+        # Alter the client's brokerclient dict to use our mocked broker
+        client.clients = mocked_brokers
+        client._collect_hosts_d = None
+        with patch.object(kclient, 'log') as klog:
+            respD = client._send_broker_unaware_request(1, 'fake request')
+            reactor.advance(client.timeout + 1)  # fire the timeout errback
+            klog.error.assert_called_once_with(
+                'Reactor was starved for %f seconds during request.', 10.0)
         self.successResultOf(
             self.failUnlessFailure(respD, KafkaUnavailableError))
         self.assertTrue(cbArg[0].check(RequestTimedOutError))
@@ -281,8 +354,7 @@ class TestKafkaClient(unittest.TestCase):
             self.successResultOf(
                 self.failUnlessFailure(d3, KafkaUnavailableError))
 
-    @patch('afkak.client.KafkaCodec')
-    def test_get_leader_for_partitions_reloads_metadata(self, kCodec):
+    def test_get_leader_for_partitions_reloads_metadata(self):
         """
         test_get_leader_for_partitions_reloads_metadata
         Get leader for partitions reload metadata if it is not available
@@ -452,7 +524,7 @@ class TestKafkaClient(unittest.TestCase):
         ]))
 
     @patch('afkak.client._get_IP_addresses')
-    def test__collect_hosts__does_not_resolve_host(self, IP_addresses):
+    def test__collect_hosts__does_not_resolve_bad_host(self, IP_addresses):
         hosts = "..."
         IP_addresses.return_value = None
         result = self.successResultOf(_collect_hosts(hosts))
@@ -504,22 +576,43 @@ class TestKafkaClient(unittest.TestCase):
         ]))
 
     @patch('afkak.client.DNSclient.lookupAddress')
-    def test__get_IP_addresses(self, RRHeader):
-        mock = MagicMock()
+    def test__get_IP_addresses_success(self, lookupAddress):
+        name = 'fully.qualified.domain.name.'
         ip_address = '127.0.0.1'
-        mock.payload.dottedQuad.return_value = ip_address
-        RRHeader.return_value = ([mock], [], [])
-        result = self.successResultOf(_get_IP_addresses('address'))
+        answer = RRHeader(
+            name=name, type=dns.A,
+            payload=Record_A(address=ip_address))
+        lookupAddress.return_value = ([answer], [], [])
+        result = self.successResultOf(_get_IP_addresses(name))
         self.assertEqual(result, [ip_address])
 
-        # Test given a host_address which is an IP address
-        result = self.successResultOf(_get_IP_addresses(ip_address))
+    def test__get_IP_addresses_addr(self):
+        ip_address = '127.0.0.1'
+        name = ip_address
+        result = self.successResultOf(_get_IP_addresses(name))
         self.assertEqual(result, [ip_address])
 
-        # Test given a bad host_address
-        RRHeader.return_value = ([], [], [])
-        result = self.successResultOf(_get_IP_addresses(' '))
-        self.assertEqual(None, result)
+    @patch('afkak.client.DNSclient.lookupAddress')
+    def test__get_IP_addresses_cname(self, lookupAddress):
+        cname = 'cname.qualified.domain.name.'
+        name = 'fully.qualified.domain.name.'
+        ip_address = '127.0.0.1'
+        cname = RRHeader(
+            name=cname, type=dns.CNAME,
+            payload=Record_CNAME(name=name))
+        answer = RRHeader(
+            name=name, type=dns.A,
+            payload=Record_A(address=ip_address))
+        lookupAddress.return_value = ([cname, answer], [], [])
+        result = self.successResultOf(_get_IP_addresses(name))
+        self.assertEqual(result, [ip_address])
+
+    @patch('afkak.client.DNSclient.lookupAddress')
+    def test__get_IP_addresses_fail(self, lookupAddress):
+        name = 'nosuch.qualified.domain.name.'
+        lookupAddress.side_effect = DNSNameError('No Such Name!')
+        result = self.successResultOf(_get_IP_addresses(name))
+        self.assertEqual(result, [])
 
     @patch('afkak.client.KafkaBrokerClient')
     def test_get_brokerclient(self, broker):
@@ -572,6 +665,7 @@ class TestKafkaClient(unittest.TestCase):
         bkr = "aBroker"
         client.reset_all_metadata = MagicMock()
         client.load_metadata_for_topics = MagicMock()
+        client._collect_hosts_d = None
         client._update_broker_state(bkr, False, e)
         client.reset_all_metadata.assert_called_once_with()
         client.load_metadata_for_topics.assert_called_once_with()
@@ -1619,3 +1713,58 @@ class TestKafkaClient(unittest.TestCase):
         self.assertTrue(
             self.failureResultOf(respD2, ConsumerCoordinatorNotAvailableError))
         client.close()
+
+    def test_client_reresolves_on_failure(self):
+        """test_client_reresolves_on_failure
+
+        Test that when the client fails to contact all brokers that it tries
+        to re-resolve the IPs of the brokers
+        """
+        mocked_brokers = {
+            ('kafka01', 9092): MagicMock(),
+        }
+
+        # inject side effects (makeRequest returns deferreds that are
+        # pre-failed with a timeout...
+        mocked_brokers[
+            ('kafka01', 9092)
+            ].makeRequest.side_effect = lambda a, b: fail(
+            RequestTimedOutError("kafka01 went away (unittest)"))
+
+        client = KafkaClient(hosts=['kafka01:9092'])
+        # Alter the client's brokerclient dict
+        client.clients = mocked_brokers
+        client._collect_hosts_d = None
+        # Get the deferred (should be already failed)
+        fail1 = client._send_broker_unaware_request(1, 'fake request')
+        # check it
+        self.successResultOf(
+            self.failUnlessFailure(fail1, KafkaUnavailableError))
+
+        # Make sure we're flagged for lookup
+        self.assertTrue(client._collect_hosts_d)
+
+        # Check that the proper calls were made
+        for key, brkr in mocked_brokers.iteritems():
+            brkr.makeRequest.assert_called_with(1, 'fake request')
+
+        # Patch the lookup and retry the request
+        with patch("afkak.client.DNSclient.lookupAddress") as lookupAddr:
+            answer = Mock(
+                **{'type': dns.A,
+                   'payload.dottedQuad.return_value': "1.2.3.4",})
+            lookupAddr.return_value = (
+                [answer], None, None)
+
+            # Patch away client._get_brokerclient. We'll end up with no brokers
+            get_broker = Mock()
+            client._get_brokerclient = get_broker
+
+            fail2 = client._send_broker_unaware_request(2, 'fake request')
+            # check it
+            self.successResultOf(
+                self.failUnlessFailure(fail2, KafkaUnavailableError))
+
+        # Check that the proper calls were made
+        get_broker.assert_called_with('1.2.3.4', 9092)
+        lookupAddr.assert_called_with('kafka01')