+++ afkak/test/test_client_integration.py
@@ -14,6 +14,7 @@ from afkak import (KafkaClient,)
 from afkak.common import (
     FetchRequest, OffsetFetchRequest, OffsetCommitRequest,
     ProduceRequest, OffsetRequest, ConsumerCoordinatorNotAvailableError,
+    NotCoordinatorForConsumerError,
     )
 from afkak.kafkacodec import (create_message)
 from fixtures import ZookeeperFixture, KafkaFixture
@@ -42,14 +43,15 @@ class TestAfkakClientIntegration(KafkaIntegrationTestCase):
         zk_chroot = random_string(10)
         replicas = 3
         partitions = 2
+        max_bytes = 12 * 1048576  # 12 MB
 
         cls.zk = ZookeeperFixture.instance()
-        kk_args = [cls.zk.host, cls.zk.port, zk_chroot, replicas, partitions]
+        kk_args = [cls.zk.host, cls.zk.port, zk_chroot, replicas, partitions, max_bytes]
         cls.kafka_brokers = [
             KafkaFixture.instance(i, *kk_args) for i in range(replicas)]
 
         hosts = ['%s:%d' % (b.host, b.port) for b in cls.kafka_brokers]
-        cls.client = KafkaClient(hosts, timeout=1500, clientId=__name__)
+        cls.client = KafkaClient(hosts, timeout=2500, clientId=__name__)
 
         # Startup the twisted reactor in a thread. We need this before the
         # the KafkaClient can work, since KafkaBrokerClient relies on the
@@ -105,11 +107,55 @@ class TestAfkakClientIntegration(KafkaIntegrationTestCase):
         self.assertEqual(produce_resp.partition, 0)
         self.assertEqual(produce_resp.offset, 0)
 
+    @kafka_versions("all")
+    @deferred(timeout=5)
+    @inlineCallbacks
+    def test_produce_large_request(self):
+        produce = ProduceRequest(
+            self.topic, 0,
+            [create_message(self.topic + " message %d: " % i
+                            + "0123456789" * 1048576)
+             for i in range(5)])
+
+        produce_resp, = yield self.client.send_produce_request([produce])
+        self.assertEqual(produce_resp.error, 0)
+        self.assertEqual(produce_resp.topic, self.topic)
+        self.assertEqual(produce_resp.partition, 0)
+        self.assertEqual(produce_resp.offset, 0)
+
+    @kafka_versions("all")
+    @deferred(timeout=5)
+    @inlineCallbacks
+    def test_roundtrip_large_request(self):
+        log.debug('Timestamp Before ProduceRequest')
+        # Single message of 5 MBish
+        produce = ProduceRequest(self.topic, 0, [create_message(
+            self.topic + " message 0: " + ("0123456789" * 10 + '\n') * 51909)])
+        log.debug('Timestamp After ProduceRequest')
+
+        produce_resp, = yield self.client.send_produce_request([produce])
+        log.debug('Timestamp After Send')
+        self.assertEqual(produce_resp.error, 0)
+        self.assertEqual(produce_resp.topic, self.topic)
+        self.assertEqual(produce_resp.partition, 0)
+        self.assertEqual(produce_resp.offset, 0)
+
+        # Fetch request with max size of 6MB
+        fetch = FetchRequest(self.topic, 0, 0, 6 * 1048576)
+        fetch_resp, = yield self.client.send_fetch_request(
+            [fetch], max_wait_time=500)
+        self.assertEqual(fetch_resp.error, 0)
+        self.assertEqual(fetch_resp.topic, self.topic)
+        self.assertEqual(fetch_resp.partition, 0)
+
+        messages = list(fetch_resp.messages)
+        self.assertEqual(len(messages), 1)
+
     ####################
     #   Offset Tests   #
     ####################
 
-    @kafka_versions("0.8.1", "0.8.1.1", "0.8.2.1", "0.9.0.1")
+    @kafka_versions("0.8.1", "0.8.1.1", "0.8.2.1", "0.8.2.2", "0.9.0.1")
     @deferred(timeout=5)
     @inlineCallbacks
     def test_send_offset_request(self):
@@ -120,7 +166,7 @@ class TestAfkakClientIntegration(KafkaIntegrationTestCase):
         self.assertEqual(resp.partition, 0)
         self.assertEqual(resp.offsets, (0,))
 
-    @kafka_versions("0.8.2.1", "0.9.0.1")
+    @kafka_versions("0.8.2.1", "0.8.2.2", "0.9.0.1")
     @deferred(timeout=15)
     @inlineCallbacks
     def test_commit_fetch_offsets(self):
@@ -151,6 +197,16 @@ class TestAfkakClientIntegration(KafkaIntegrationTestCase):
                     c_group, attempt)
                 time.sleep(0.5)
                 continue
+            except NotCoordinatorForConsumerError:  # pragma: no cover
+                # Kafka seems to have a timing issue: If we ask broker 'A' who
+                # the ConsumerCoordinator is for a auto-created, not extant
+                # topic, the assigned broker may not realize it's been so
+                # designated by the time we find out and make our request.
+                log.info(
+                    "Coordinator is not coordinator!!: %s Attempt: %d of 20",
+                    c_group, attempt)
+                time.sleep(0.5)
+                continue
             break
         self.assertEqual(getattr(resp, 'error', -1), 0)
 
@@ -158,6 +214,6 @@ class TestAfkakClientIntegration(KafkaIntegrationTestCase):
         (resp,) = yield self.client.send_offset_fetch_request(c_group, [req])
         self.assertEqual(resp.error, 0)
         self.assertEqual(resp.offset, offset)
-        # broker doesn't seem to return proper metadata
+        # Check we received the proper metadata in the response
         self.assertEqual(resp.metadata, metadata)
         log.debug("test_commit_fetch_offsets: Test Complete.")